+++ lib/nepali_utils.dart
@@ -1,6 +1,6 @@
-export 'src/date_converter.dart';
 export 'src/nepali_date_time.dart';
-export 'src/nepali_date_formatter.dart';
-export 'src/nepali_number.dart';
+export 'src/nepali_date_format.dart';
+export 'src/nepali_number_format.dart';
 export 'src/nepali_unicode.dart';
 export 'src/nepali_moment.dart';
+export 'src/nepali_language.dart';