+++ lib/src/nepali_date_format.dart
@@ -1,25 +1,25 @@
 import 'nepali_date_time.dart';
-import 'nepali_number.dart';
+import 'nepali_language.dart';
+import 'nepali_unicode.dart';
 
-enum Language { ENGLISH, NEPALI }
-
-class NepaliDateFormatter {
-  final String pattern;
-  final Language language;
+class NepaliDateFormat {
+  Language _language;
   String _pattern;
   String _checkPattern;
   bool _firstRun = true;
   int _index = 0;
 
-  NepaliDateFormatter(
-    this.pattern, {
-    this.language = Language.ENGLISH,
-  });
+  NepaliDateFormat(
+    String pattern, {
+    Language language,
+  }) {
+    _pattern = pattern;
+    _language = language ?? Language.ENGLISH;
+  }
 
   String format(NepaliDateTime dateTime) {
-    //print(_checkPattern);
     if (_firstRun) {
-      _checkPattern = _pattern = pattern;
+      _checkPattern = _pattern;
       _firstRun = false;
     }
     for (int i = 0; i < _matchers.length; i++) {
@@ -68,35 +68,35 @@ class NepaliDateFormatter {
   void _format(String match, NepaliDateTime dateTime) {
     switch (match) {
       case 'G':
-        _replacer(match, language == Language.ENGLISH ? 'BS' : 'बि सं');
+        _replacer(match, _language == Language.ENGLISH ? 'BS' : 'बि सं');
         break;
       case 'GG':
-        _replacer(match, language == Language.ENGLISH ? 'B.S.' : 'बि.सं.');
+        _replacer(match, _language == Language.ENGLISH ? 'B.S.' : 'बि.सं.');
         break;
       case 'GGG':
         _replacer(match,
-            language == Language.ENGLISH ? 'Bikram Sambat' : 'बिक्रम संबत');
+            _language == Language.ENGLISH ? 'Bikram Sambat' : 'बिक्रम संबत');
         break;
       case 'y':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.year}'
-                : '${NepaliNumber.from(dateTime.year)}');
+                : '${NepaliUnicode.convert('${dateTime.year}')}');
         break;
       case 'yy':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.year.toString().substring(2)}'
-                : '${NepaliNumber.fromString(dateTime.year.toString().substring(2))}');
+                : '${NepaliUnicode.convert(dateTime.year.toString().substring(2))}');
         break;
       case 'yyyy':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.year}'
-                : '${NepaliNumber.from(dateTime.year)}');
+                : '${NepaliUnicode.convert('${dateTime.year}')}');
         break;
       case 'Q':
         _replacer(match, '${_getQuarter(dateTime.month)}');
@@ -114,9 +114,9 @@ class NepaliDateFormatter {
       case 'M':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.month}'
-                : '${NepaliNumber.from(dateTime.month)}');
+                : '${NepaliUnicode.convert('${dateTime.month}')}');
         break;
       case 'MM':
         _replacer(match, _prependZero(dateTime.month));
@@ -130,9 +130,9 @@ class NepaliDateFormatter {
       case 'd':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.day}'
-                : '${NepaliNumber.from(dateTime.day)}');
+                : '${NepaliUnicode.convert('${dateTime.day}')}');
         break;
       case 'dd':
         _replacer(match, _prependZero(dateTime.day));
@@ -143,7 +143,7 @@ class NepaliDateFormatter {
       case 'EE':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? _weekDayString(dateTime.weekDay).substring(0, 3)
                 : _weekDayString(dateTime.weekDay).replaceFirst('बार', ''));
         break;
@@ -153,7 +153,7 @@ class NepaliDateFormatter {
       case 'a':
         _replacer(
           match,
-          language == Language.ENGLISH
+          _language == Language.ENGLISH
               ? dateTime.hour >= 12 ? 'pm' : 'am'
               : dateTime.hour < 12
                   ? 'बिहान'
@@ -167,7 +167,7 @@ class NepaliDateFormatter {
       case 'aa':
         _replacer(
           match,
-          language == Language.ENGLISH
+          _language == Language.ENGLISH
               ? dateTime.hour >= 12 ? 'PM' : 'AM'
               : dateTime.hour < 12
                   ? 'बिहान'
@@ -187,9 +187,9 @@ class NepaliDateFormatter {
       case 'H':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.hour}'
-                : NepaliNumber.from(dateTime.hour));
+                : NepaliUnicode.convert('${dateTime.hour}'));
         break;
       case 'HH':
         _replacer(match, _prependZero(dateTime.hour));
@@ -197,9 +197,9 @@ class NepaliDateFormatter {
       case 'm':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.minute}'
-                : NepaliNumber.from(dateTime.minute));
+                : NepaliUnicode.convert('${dateTime.minute}'));
         break;
       case 'mm':
         _replacer(match, _prependZero(dateTime.minute));
@@ -207,9 +207,9 @@ class NepaliDateFormatter {
       case 's':
         _replacer(
             match,
-            language == Language.ENGLISH
+            _language == Language.ENGLISH
                 ? '${dateTime.second}'
-                : NepaliNumber.from(dateTime.second));
+                : NepaliUnicode.convert('${dateTime.second}'));
         break;
       case 'ss':
         _replacer(match, _prependZero(dateTime.second));
@@ -250,39 +250,43 @@ class NepaliDateFormatter {
   String _threeDigitMaker(int number) {
     String numString = number.toString();
     if (numString.length == 1) {
-      return language == Language.ENGLISH
+      return _language == Language.ENGLISH
           ? '00$numString'
-          : '००${NepaliNumber.from(number)}';
+          : '००${NepaliUnicode.convert('$number')}';
     }
     if (numString.length == 2) {
-      return language == Language.ENGLISH
+      return _language == Language.ENGLISH
           ? '0$numString'
-          : '०${NepaliNumber.from(number)}';
+          : '०${NepaliUnicode.convert('$number')}';
     }
-    return language == Language.ENGLISH
+    return _language == Language.ENGLISH
         ? numString.substring(0, 3)
-        : NepaliNumber.fromString(numString.substring(0, 3));
+        : NepaliUnicode.convert(numString.substring(0, 3));
   }
 
   String _clockHour(int hour, {bool prependZero = false}) {
     if (hour > 12) {
-      return language == Language.ENGLISH
+      return _language == Language.ENGLISH
           ? '${hour - 12}'
-          : '${NepaliNumber.from(hour - 12)}';
+          : '${NepaliUnicode.convert('${hour - 12}')}';
     } else if (hour == 12) {
-      return language == Language.ENGLISH ? '12' : '१२';
+      return _language == Language.ENGLISH ? '12' : '१२';
     } else {
-      return language == Language.ENGLISH
+      return _language == Language.ENGLISH
           ? prependZero ? _prependZero(hour) : '${hour}'
-          : prependZero ? _prependZero(hour) : '${NepaliNumber.from(hour)}';
+          : prependZero
+              ? _prependZero(hour)
+              : '${NepaliUnicode.convert('$hour')}';
     }
   }
 
   String _prependZero(int number) => number < 10
-      ? language == Language.ENGLISH
+      ? _language == Language.ENGLISH
           ? '0$number'
-          : '०${NepaliNumber.from(number)}'
-      : language == Language.ENGLISH ? '$number' : NepaliNumber.from(number);
+          : '०${NepaliUnicode.convert('$number')}'
+      : _language == Language.ENGLISH
+          ? '$number'
+          : NepaliUnicode.convert('$number');
 
   void _replacer(String match, String replaceWith) {
     _pattern = _pattern.replaceFirst(match, replaceWith, _index);
@@ -292,31 +296,31 @@ class NepaliDateFormatter {
   String _weekDayString(int day, {bool short = false}) {
     switch (day) {
       case 1:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'S' : 'Sunday'
             : short ? 'आ' : 'आइतबार';
       case 2:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'M' : 'Monday'
             : short ? 'सो' : 'सोमबार';
       case 3:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'T' : 'Tuesday'
             : short ? 'मं' : 'मंगलबार';
       case 4:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'W' : 'Wednesday'
             : short ? 'बु' : 'बुधबार';
       case 5:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'T' : 'Thursday'
             : short ? 'बि' : 'बिहिबार';
       case 6:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'F' : 'Friday'
             : short ? 'शु' : 'शुक्रबार';
       case 7:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'S' : 'Saturday'
             : short ? 'श' : 'शनिबार';
       default:
@@ -327,51 +331,51 @@ class NepaliDateFormatter {
   String _monthString(int month, {short = false}) {
     switch (month) {
       case 1:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Bai' : 'Baishakh'
             : short ? 'बै' : 'बैशाख';
       case 2:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Jes' : 'Jestha'
             : short ? 'जे' : 'जेष्ठ';
       case 3:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Asa' : 'Ashadh'
             : short ? 'अ' : 'अषाढ';
       case 4:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Shr' : 'Shrawan'
             : short ? 'श्रा' : 'श्रावण';
       case 5:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Bha' : 'Bhadra'
             : short ? 'भा' : 'भाद्र';
       case 6:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Ash' : 'Ashwin'
             : short ? 'आ' : 'आश्विन';
       case 7:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Kar' : 'Kartik'
             : short ? 'का' : 'कार्तिक';
       case 8:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Marg' : 'Mangsir'
             : short ? 'मं' : 'मंसिर';
       case 9:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Pou' : 'Poush'
             : short ? 'पौ' : 'पौष';
       case 10:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Mag' : 'Magh'
             : short ? 'मा' : 'माघ';
       case 11:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Fal' : 'Falgun'
             : short ? 'फा' : 'फाल्गुन';
       case 12:
-        return language == Language.ENGLISH
+        return _language == Language.ENGLISH
             ? short ? 'Cha' : 'Chaitra'
             : short ? 'चै' : 'चैत्र';
       default:
@@ -400,4 +404,70 @@ class NepaliDateFormatter {
     // e.g. in "hh:mm:ss" will match the colons.
     RegExp("^[^\'GyMkSEahKHcLQdDmsvzZ]+")
   ];
+
+  NepaliDateFormat.d([Language language]) : this("d", language: language);
+  NepaliDateFormat.E([Language language]) : this("EE", language: language);
+  NepaliDateFormat.EEEE([Language language]) : this("EEE", language: language);
+  NepaliDateFormat.LLL([Language language]) : this("MMM", language: language);
+  NepaliDateFormat.LLLL([Language language]) : this("MMMM", language: language);
+  NepaliDateFormat.M([Language language]) : this("M", language: language);
+  NepaliDateFormat.Md([Language language]) : this("M/d", language: language);
+  NepaliDateFormat.MEd([Language language])
+      : this("EE, M/d", language: language);
+  NepaliDateFormat.MMM([Language language]) : this("MMM", language: language);
+  NepaliDateFormat.MMMd([Language language])
+      : this("MMM d", language: language);
+  NepaliDateFormat.MMMEd([Language language])
+      : this("EEE, MMM d", language: language);
+  NepaliDateFormat.MMMM([Language language]) : this("MMMM", language: language);
+  NepaliDateFormat.MMMMd([Language language])
+      : this("MMMM d", language: language);
+  NepaliDateFormat.MMMMEEEEd([Language language])
+      : this("EEE, MMMM d", language: language);
+  NepaliDateFormat.QQQ([Language language]) : this("QQQ", language: language);
+  NepaliDateFormat.QQQQ([Language language]) : this("QQQQ", language: language);
+  NepaliDateFormat.y([Language language]) : this("y", language: language);
+  NepaliDateFormat.yM([Language language]) : this("y/MM", language: language);
+  NepaliDateFormat.yMd([Language language])
+      : this("y/MM/dd", language: language);
+  NepaliDateFormat.yMEd([Language language])
+      : this("EE, y/MM/dd", language: language);
+  NepaliDateFormat.yMMM([Language language])
+      : this("MMM y", language: language);
+  NepaliDateFormat.yMMMd([Language language])
+      : this("MMM d, y", language: language);
+  NepaliDateFormat.yMMMEd([Language language])
+      : this("EE, MMM d, y", language: language);
+  NepaliDateFormat.yMMMM([Language language])
+      : this("MMMM y", language: language);
+  NepaliDateFormat.yMMMMd([Language language])
+      : this("MMMM d, y", language: language);
+  NepaliDateFormat.yMMMMEEEEd([Language language])
+      : this("EEE, MMMM d, y", language: language);
+  NepaliDateFormat.yQQQ([Language language])
+      : this("QQQ y", language: language);
+  NepaliDateFormat.yQQQQ([Language language])
+      : this("QQQQ y", language: language);
+  NepaliDateFormat.H([Language language]) : this("H", language: language);
+  NepaliDateFormat.Hm([Language language]) : this("HH:MM", language: language);
+  NepaliDateFormat.Hms([Language language])
+      : this("HH:mm:ss", language: language);
+  NepaliDateFormat.j([Language language])
+      : this(
+          language == Language.NEPALI ? "aa h" : "h aa",
+          language: language,
+        );
+  NepaliDateFormat.jm([Language language])
+      : this(
+          language == Language.NEPALI ? "aa h:mm" : "h:mm aa",
+          language: language,
+        );
+  NepaliDateFormat.jms([Language language])
+      : this(
+          language == Language.NEPALI ? "aa h:mm:ss" : "h:mm:ss aa",
+          language: language,
+        );
+  NepaliDateFormat.m([Language language]) : this("h", language: language);
+  NepaliDateFormat.ms([Language language]) : this("hh:mm", language: language);
+  NepaliDateFormat.s([Language language]) : this("s", language: language);
 }