+++ lib/src/nepali_language.dart
@@ -1,0 +1,1 @@
+enum Language { ENGLISH, NEPALI }